+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Introduce local interfaces to make easier to test the organization handler.
+
 ## [0.8.0] - 2021-05-24
 
 ### Changed