+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `apiextensions` dependency to v3.
 - Updated `operatorkit` dependency to v5.
 - Updated `k8sclient` dependency to v5.
+- Always try to create organization in `companyd` even when organization namespace already exists.
 
 ## [0.8.0] - 2021-05-24