+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `operatorkit` dependency to v5.
 - Updated `k8sclient` dependency to v5.
 - Always try to create organization in `companyd` even when organization namespace already exists.
+- Pin `jwt-go` dependency to avoid version with vulnerabilities.
 
 ## [0.8.0] - 2021-05-24